Create shared component for row overlay
