+++ common/v2/components/AccountList.tsx
@@ -5,7 +5,7 @@ import { Button, Copyable, Identicon } from '@mycrypto/ui';
 
 import { translateRaw } from 'v2/translations';
 import { ROUTE_PATHS, Fiats } from 'v2/config';
-import { CollapsibleTable, Network } from 'v2/components';
+import { CollapsibleTable, Network, RowDeleteOverlay } from 'v2/components';
 import { default as Typography } from 'v2/components/Typography'; // @TODO solve Circular Dependency issue
 import { truncate, IS_MOBILE } from 'v2/utils';
 import { BREAK_POINTS, COLORS, breakpointToNumber } from 'v2/theme';
@@ -90,39 +90,6 @@ const TableContainer = styled.div`
   overflow: auto;
 `;
 
-const TableOverlay = styled.div`
-  height: 67px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  background-color: #b5bfc7;
-  color: #fff;
-  padding: 1em;
-`;
-
-const OverlayText = styled(Typography)`
-  color: #fff;
-  flex-grow: 1;
-  text-overflow: hidden;
-  width: 50%;
-`;
-
-const OverlayButtons = styled.div`
-  align-self: flex-end;
-`;
-
-const OverlayDelete = styled(Button)`
-  font-size: 14px;
-  margin-left: 5px;
-`;
-
-const OverlayCancel = styled(Button)`
-  font-size: 14px;
-  padding-top: 8px;
-  padding-bottom: 8px;
-  margin-left: 5px;
-`;
-
 type DeleteAccount = (uuid: string) => void;
 type UpdateAccount = (uuid: string, accountData: ExtendedAccount) => void;
 interface AccountListProps {
@@ -203,28 +170,19 @@ function buildAccountTable(
     head: deletable ? [...columns, translateRaw('ACCOUNT_LIST_DELETE')] : columns,
     overlay:
       overlayRows && overlayRows[0] !== undefined ? (
-        <TableOverlay>
-          <OverlayText>
-            Are you sure you want to delete{' '}
-            {getLabelByAccount(accounts[overlayRows[0]], addressBook) !== undefined
-              ? getLabelByAccount(accounts[overlayRows[0]], addressBook)!.label
-              : ''}{' '}
-            account with address: {truncate(accounts[overlayRows[0]].address)}?
-          </OverlayText>
-          <OverlayButtons>
-            <OverlayDelete
-              onClick={() => {
-                deleteAccount(accounts[overlayRows[0]].uuid);
-                setDeletingIndex(undefined);
-              }}
-            >
-              Delete
-            </OverlayDelete>
-            <OverlayCancel secondary={true} onClick={() => setDeletingIndex(undefined)}>
-              Cancel
-            </OverlayCancel>
-          </OverlayButtons>
-        </TableOverlay>
+        <RowDeleteOverlay
+          prompt={`Are you sure you want to delete 
+              ${
+                getLabelByAccount(accounts[overlayRows[0]], addressBook) !== undefined
+                  ? getLabelByAccount(accounts[overlayRows[0]], addressBook)!.label
+                  : ''
+              } account with address: ${truncate(accounts[overlayRows[0]].address)} ?`}
+          deleteAction={() => {
+            deleteAccount(accounts[overlayRows[0]].uuid);
+            setDeletingIndex(undefined);
+          }}
+          cancelAction={() => setDeletingIndex(undefined)}
+        />
       ) : (
         <></>
       ),
